Separating node topology, racks and failure domains logic
Taking out the node topology, racks and failure domains' logic to a
separate file. Thus cleaning up some more clutter from 'reconcile.go'
file.
